refactor(behavior-analytics): one class per coordinate system
The trajectory and state-management hierarchies described distinctions that did
not exist. TrajectoryI was a bare alias of TrajectoryBase; StateMgmtE and
StateMgmtI each held a single method that only picked a trajectory type. The
unsuffixed name meant geographic in both, so `Trajectory` and `StateMgmt` read as
defaults while actually being the least common case.

Collapse each to one class per coordinate system, with the geographic variant
carrying the suffix it always should have:

- TrajectoryBase + TrajectoryE + TrajectoryI -> Trajectory, plus TrajectoryG.
  The raw computations become _raw_* helpers so both the calibration-type
  dispatch and TrajectoryG call them directly; previously TrajectoryG reached
  upward with super(), which after a merge would have silently double-converted
  every geographic speed.
- StateMgmtBase + StateMgmtE + StateMgmtI -> StateMgmt, plus StateMgmtG.
- Calibration -> CalibrationG, so no calibration class holds the unsuffixed name.
  CalibrationBase stays: its three variants are real implementations, not
  pass-throughs, so there is nothing to collapse.

Extract the point series sampling into point_sampling.py. It is a self-contained
mechanism -- keep 1 point in N, double N when the series outgrows
behaviorMaxPoints -- with three invariants that were previously stated only in
inline comments. Verified behaviour-identical against the original inline block
over 5204 randomised batches.

The parity correction that keeps the 1-in-N stride continuous across halving was
untested: both existing tests derived the expected result from the phase the code
had just produced, so they could not fail. They now assert against the original
message stream, and removing the correction fails five tests.
