+++ services/analytics/behavior-analytics/src/mdx/analytics/core/schema/trajectory/trajectory.py
@@ -13,298 +13,367 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import logging
+import math
+from datetime import datetime
 from functools import cached_property
 
-from pydantic import computed_field
+import numpy as np
+from pydantic import BaseModel, computed_field
 
-from mdx.analytics.core.schema.models import Coordinate, Location
-from mdx.analytics.core.schema.trajectory.trajectory_base import TrajectoryBase
-from mdx.analytics.core.utils.distance_util import MPS_TO_MPH, bearing, haversine_distance_coords
+from mdx.analytics.core.schema.models import Coordinate, GeoLocation, Point
+from mdx.analytics.core.utils.distance_util import MPS_TO_MPH, euclidean_distance
+from mdx.analytics.core.transform.calibration.calibration_base import CalibrationType
+
+logger = logging.getLogger(__name__)
 
 
-class Trajectory(TrajectoryBase):
+class Trajectory(BaseModel):
     """
-    Geographic trajectory representation with configurable direction modes.
+    Base class for trajectory representation and analysis.
 
-    This class extends TrajectoryBase to provide trajectory analysis in geographic space
-    with configurable direction modes and clustering options. It supports both Euclidean
-    and geographic coordinate systems with customizable direction representations.
+    This class provides fundamental functionality for working with trajectories, including:
+    - Point storage and access
+    - Trajectory smoothing
+    - Distance calculations
+    - Speed calculations
+    - Direction and bearing calculations
+    - Geographic location representation
 
     :ivar str id: Unique identifier for the trajectory
     :ivar datetime start: Start time of the trajectory
     :ivar datetime end: End time of the trajectory
     :ivar list[Coordinate] points: List of coordinate points forming the trajectory
-    :ivar bool enable_geo: Enable geographic coordinate system (default: False)
-    :ivar int direction_mode: Direction representation mode (default: 0)
-        - 0: Basic cardinal directions (N, E, S, W)
-        - 1: Intercardinal directions (N, NE, E, SE, S, SW, W, NW)
-        - 2: Secondary intercardinal directions (16 directions)
-    :ivar int direction_based_cluster_mode: Direction clustering mode (default: 1)
-        - 0: 4-direction clustering
-        - 1: 8-direction clustering
-        - 2: 16-direction clustering
     :ivar int smooth_min_points: Minimum number of points required for computing smoothed trajectory and speed over time (default: 20)
-    :ivar int smooth_window_size: Window size for smoothing (default: 5)
+    :ivar int smooth_window_size: Window size for smoothing  (default: 5)
     :ivar int distance_stride: Stride for distance calculation (default: 5)
     :ivar int speed_segment_size: Size of speed segment for speed over time calculation (default: 10)
 
     Examples::
 
-        # Create a geographic trajectory with different direction modes
+        # Basic trajectory creation
         from datetime import datetime
         from mdx.analytics.core.schema.models import Coordinate
 
-        # Create a trajectory moving northeast
-        points = [
-            Coordinate(x=-74.0060, y=40.7128, z=0),  # New York
-            Coordinate(x=-73.9352, y=40.7306, z=0),  # Brooklyn
-            Coordinate(x=-73.7781, y=40.6413, z=0)   # JFK Airport
-        ]
-
-        # Basic cardinal directions (N, E, S, W)
-        trajectory_basic = Trajectory(
-            id="traj_geo1",
+        # Create a simple trajectory with 3 points
+        points = [Coordinate(x=0, y=0, z=0), Coordinate(x=1, y=1, z=0)]
+        trajectory = Trajectory(
+            id="traj1",
             start=datetime.now(),
             end=datetime.now(),
-            points=points,
-            enable_geo=True,
-            direction_mode=0
+            points=points
         )
-        print(f"Basic direction: {trajectory_basic.direction}")
 
-        # Intercardinal directions (N, NE, E, SE, S, SW, W, NW)
-        trajectory_inter = Trajectory(
-            id="traj_geo2",
-            start=datetime.now(),
-            end=datetime.now(),
-            points=points,
-            enable_geo=True,
-            direction_mode=1
-        )
-        print(f"Intercardinal direction: {trajectory_inter.direction}")
+        # Access trajectory properties
+        print(f"Total distance: {trajectory.distance} meters")
+        print(f"Average speed: {trajectory.speed} m/s")
+        print(f"Direction: {trajectory.direction}")
+        print(f"Bearing: {trajectory.bearing} degrees")
+
+        # Get smoothed trajectory
+        smoothed_points = trajectory.smooth_trajectory
+
+        # Get speed over time segments
+        speeds = trajectory.speed_over_time
     """
 
-    enable_geo: bool = False
-    direction_mode: int = 0
-    direction_based_cluster_mode: int = 1
+    id: str
+    start: datetime
+    end: datetime
+    points: list[Coordinate]
+    smooth_min_points: int = 20
+    smooth_window_size: int = 5
+    distance_stride: int = 5
+    speed_segment_size: int = 10
 
     @computed_field
     @cached_property
-    def bearing(self) -> float:
+    def head(self) -> Coordinate:
         """
-        Calculate the bearing of the trajectory, supporting both geographic and Euclidean systems.
+        Get the first point of the trajectory.
 
-        :return float: Bearing in degrees (0-360)
-
-        Example::
-
-            # Geographic bearing (using latitude/longitude)
-            points = [
-                Coordinate(x=-74.0060, y=40.7128, z=0),  # New York
-                Coordinate(x=-73.9352, y=40.7306, z=0)   # Brooklyn
-            ]
-            geo_trajectory = Trajectory(
-                id="geo1",
-                start=datetime.now(),
-                end=datetime.now(),
-                points=points,
-                enable_geo=True
-            )
-            geo_bearing = geo_trajectory.bearing  # Returns geographic bearing
+        :return Coordinate: First coordinate point
         """
-        if self.enable_geo:
-            from_locaton = Location(lat=self.head.y, lon=self.head.x)
-            to_locaton = Location(lat=self.last.y, lon=self.last.x)
-            return bearing(from_locaton, to_locaton)
-        return super().bearing
+        return self.points[0]
 
     @computed_field
     @cached_property
-    def direction(self) -> str:
+    def last(self) -> Coordinate:
         """
-        Get the direction of the trajectory based on the configured direction mode.
+        Get the last point of the trajectory.
 
-        :return str: Direction string based on the current direction_mode
+        :return Coordinate: Last coordinate point
+        """
+        return self.points[-1]
+
+    @computed_field
+    @cached_property
+    def smooth_trajectory(self) -> list[Coordinate]:
+        """
+        Get a smoothed version of the trajectory using a moving average window.
+
+        :return list[Coordinate]: List of smoothed coordinates
 
         Example::
 
-            # Basic cardinal directions (N, E, S, W)
+            # Create a trajectory with some noise
             points = [
                 Coordinate(x=0, y=0, z=0),
-                Coordinate(x=1, y=1, z=0)  # Northeast
+                Coordinate(x=1, y=1.1, z=0),  # Slight noise
+                Coordinate(x=2, y=2, z=0),
+                Coordinate(x=3, y=2.9, z=0),  # Slight noise
+                Coordinate(x=4, y=4, z=0)
             ]
-            trajectory_basic = Trajectory(
-                id="basic1",
+            trajectory = Trajectory(
+                id="1",
                 start=datetime.now(),
                 end=datetime.now(),
                 points=points,
-                enable_geo=True,
-                direction_mode=0
+                moving_avg_window=3
             )
-            basic_dir = trajectory_basic.direction  # Returns "NE"
 
-            # Intercardinal directions
-            trajectory_inter = Trajectory(
-                id="inter1",
-                start=datetime.now(),
-                end=datetime.now(),
-                points=points,
-                enable_geo=True,
-                direction_mode=1
-            )
-            inter_dir = trajectory_inter.direction  # Returns "NE"
-
-            # Secondary intercardinal directions
-            trajectory_sec = Trajectory(
-                id="sec1",
-                start=datetime.now(),
-                end=datetime.now(),
-                points=points,
-                enable_geo=True,
-                direction_mode=2
-            )
-            sec_dir = trajectory_sec.direction  # Returns "ENE"
+            # Get smoothed points
+            smoothed = trajectory.smooth_trajectory
+            # The y-coordinates will be averaged over the window size
+            # reducing the effect of noise
         """
-        dirs = ["Right", "Up", "Left", "Down"]
-        if self.enable_geo:
-            if self.direction_mode == 0:
-                dirs = ["N", "E", "S", "W"]
-            elif self.direction_mode == 1:
-                dirs = ["N", "NE", "E", "SE", "S", "SW", "W", "NW"]
-            elif self.direction_mode == 2:
-                dirs = [
-                    "N",
-                    "NNE",
-                    "NE",
-                    "ENE",
-                    "E",
-                    "ESE",
-                    "SE",
-                    "SSE",
-                    "S",
-                    "SSW",
-                    "SW",
-                    "WSW",
-                    "W",
-                    "WNW",
-                    "NW",
-                    "NNW",
-                ]
-        return dirs[self.direction_index]
+        if len(self.points) < self.smooth_min_points:
+            return self.points
+
+        # Convert points to numpy arrays for vectorized operations
+        points_array = np.array([[p.x, p.y, p.z] for p in self.points])
+
+        # Create a padded array to handle the window at the start
+        padded = np.pad(points_array, ((self.smooth_window_size - 1, 0), (0, 0)), mode='edge')
+
+        # Calculate cumulative sum for efficient moving average
+        cumsum = np.cumsum(padded, axis=0)
+
+        # Calculate moving average using cumulative sum
+        # Ensure both arrays have the same shape by adjusting the slicing
+        window_size = self.smooth_window_size
+        smoothed = (cumsum[window_size:] - cumsum[:-window_size]) / window_size
+
+        # Convert back to Coordinate objects
+        return [Coordinate(x=float(x), y=float(y), z=float(z)) for x, y, z in smoothed]
 
     @computed_field
     @cached_property
-    def direction_index(self) -> int:
+    def distance(self) -> float:
         """
-        Get the index of the direction in the current direction mode's direction list.
+        Calculate the total distance of the smoothed trajectory.
 
-        :return int: Index of the direction in the current direction mode's list
+        :return float: Distance in meters
 
         Example::
 
-            # Get direction index for different modes
+            # Create a trajectory with known distance
             points = [
                 Coordinate(x=0, y=0, z=0),
-                Coordinate(x=1, y=1, z=0)  # Northeast
+                Coordinate(x=3, y=4, z=0)  # 5 meters from origin (3-4-5 triangle)
             ]
-
-            # Basic mode (4 directions)
-            trajectory_basic = Trajectory(
-                id="basic1",
+            trajectory = Trajectory(
+                id="1",
                 start=datetime.now(),
                 end=datetime.now(),
-                points=points,
-                enable_geo=True,
-                direction_mode=0
+                points=points
             )
-            basic_idx = trajectory_basic.direction_index  # Returns 1 (E)
-
-            # Intercardinal mode (8 directions)
-            trajectory_inter = Trajectory(
-                id="inter1",
-                start=datetime.now(),
-                end=datetime.now(),
-                points=points,
-                enable_geo=True,
-                direction_mode=1
-            )
-            inter_idx = trajectory_inter.direction_index  # Returns 2 (NE)
+            total_distance = trajectory.distance  # Returns approximately 5.0
         """
-        if self.enable_geo:
-            if self.direction_mode == 0:
-                return int(((self.bearing + 45) % 360) // 90)
-            if self.direction_mode == 1:
-                return int(((self.bearing + 22.5) % 360) // 45)
-            if self.direction_mode == 2:
-                return int(((self.bearing + 11.25) % 360) // 22.5)
-        return super().direction_index
+        tr = [self.smooth_trajectory[i] for i in range(0, len(self.smooth_trajectory), self.distance_stride)] + [
+            self.last
+        ]
+
+        if len(tr) < 2:
+            return self.linear_distance
+        return sum(self._calculate_distance(tr[i], tr[i + 1]) for i in range(len(tr) - 1))
 
     @computed_field
     @cached_property
-    def direction_based_cluster_index(self) -> int:
+    def linear_distance(self) -> float:
         """
-        Get the index for direction-based clustering based on the current clustering mode.
+        Calculate the linear distance between the first and last points of the smoothed trajectory.
 
-        :return int: Index for direction-based clustering
+        :return float: Linear distance in meter
+        """
+        p1 = self.smooth_trajectory[0]
+        p2 = self.smooth_trajectory[-1]
+        if len(self.smooth_trajectory) < 2:
+            return 0
+        return self._calculate_distance(p1, p2)
+
+    def _raw_speed(self) -> float:
+        """
+        Calculate the average speed of the trajectory.
+
+        :return float: Average speed in meters per second
 
         Example::
 
-            # Get clustering index for different modes
+            from datetime import datetime, timedelta
+
+            # Create a trajectory that moves 10 meters in 2 seconds
             points = [
                 Coordinate(x=0, y=0, z=0),
-                Coordinate(x=1, y=1, z=0)  # Northeast
+                Coordinate(x=6, y=8, z=0)  # 10 meters from origin
             ]
+            start = datetime.now()
+            end = start + timedelta(seconds=2)
 
-            # 4-direction clustering
-            trajectory_4 = Trajectory(
-                id="cluster4",
-                start=datetime.now(),
-                end=datetime.now(),
-                points=points,
-                enable_geo=True,
-                direction_based_cluster_mode=0
+            trajectory = Trajectory(
+                id="1",
+                start=start,
+                end=end,
+                points=points
             )
-            idx_4 = trajectory_4.direction_based_cluster_index  # Returns 1 (E)
 
-            # 8-direction clustering
-            trajectory_8 = Trajectory(
-                id="cluster8",
-                start=datetime.now(),
-                end=datetime.now(),
-                points=points,
-                enable_geo=True,
-                direction_based_cluster_mode=1
-            )
-            idx_8 = trajectory_8.direction_based_cluster_index  # Returns 2 (NE)
+            avg_speed = trajectory.speed  # Returns 5.0 m/s (10 meters / 2 seconds)
         """
-        if self.enable_geo:
-            if self.direction_based_cluster_mode == 0:
-                return int(((self.bearing + 45) % 360) // 90)
-            if self.direction_based_cluster_mode == 1:
-                return int(((self.bearing + 22.5) % 360) // 45)
-            if self.direction_based_cluster_mode == 2:
-                return int(((self.bearing + 11.25) % 360) // 22.5)
-        return super().direction_index
+        if self.time_interval == 0:
+            if len(self.points) >= 2:
+                logger.warning(f"Time interval is 0 for trajectory {self.id} with {len(self.points)} points")
+            return 0
+        return self.distance / self.time_interval
 
-    @computed_field
-    @cached_property
-    def speed(self) -> float:
-        """
-        Calculate the average speed of the trajectory
-
-        :return float: Average speed in mph
-        """
-        return super().speed * MPS_TO_MPH
-
-    @computed_field
-    @cached_property
-    def speed_over_time(self) -> list[float]:
+    def _raw_speed_over_time(self) -> list[float]:
         """
         Calculate the speed of the trajectory over time in segments.
 
         :return list[float]: List of speeds for each segment
         """
-        return [MPS_TO_MPH * x for x in super().speed_over_time]
+        if self.time_interval == 0:
+            return [self.speed]
+
+        if len(self.smooth_trajectory) < self.smooth_min_points:
+            return [self.speed]
+
+        segment_size = self.speed_segment_size
+        slices = [
+            self.smooth_trajectory[i : i + segment_size] for i in range(0, len(self.smooth_trajectory), segment_size)
+        ]
+        last_segment = self.smooth_trajectory[-segment_size:]
+        interval = self.time_interval * (segment_size - 1) / (len(self.smooth_trajectory) - 1)
+
+        trimmed_slices = slices[:-1]
+        trimmed_slices_speeds = [
+            sum(self._calculate_distance(x[i], x[i + 1]) for i in range(len(x) - 1)) / interval for x in trimmed_slices
+        ]
+
+        last_segment_speed = (
+            sum(self._calculate_distance(last_segment[i], last_segment[i + 1]) for i in range(len(last_segment) - 1))
+            / interval
+        )
+
+        speeds_over_time = trimmed_slices_speeds + [last_segment_speed]
+        speeds_over_time = [
+            (speeds_over_time[i] + speeds_over_time[i + 1]) / 2 for i in range(len(speeds_over_time) - 1)
+        ]
+
+        return [round(x, 2) for x in speeds_over_time]
+
+    def _raw_bearing(self) -> float:
+        """
+        Calculate the bearing of the trajectory from the first to the last point, with y-axis inverted.
+
+        This uses image coordinate convention where y increases downward.
+
+        :return float: Bearing in degrees (0-360)
+
+        Example::
+
+            # Create a trajectory moving right in image coordinates
+            points = [
+                Coordinate(x=0, y=0, z=0),
+                Coordinate(x=1, y=0, z=0)
+            ]
+            trajectory = Trajectory(
+                id="1",
+                start=datetime.now(),
+                end=datetime.now(),
+                points=points
+            )
+
+            angle = trajectory.bearing  # Returns 0.0 degrees (right)
+        """
+        brng = math.atan2(-self.last.y + self.head.y, self.last.x - self.head.x) * 180 / math.pi
+        return (brng + 360) % 360
+
+    @computed_field
+    @cached_property
+    def direction(self) -> str:
+        """
+        Get the cardinal direction of the trajectory.
+
+        :return str: Cardinal direction ("Right", "Up", "Left", "Down")
+
+        Example::
+
+            # Create a trajectory moving upward
+            points = [
+                Coordinate(x=0, y=0, z=0),
+                Coordinate(x=0, y=1, z=0)
+            ]
+            trajectory = Trajectory(
+                id="1",
+                start=datetime.now(),
+                end=datetime.now(),
+                points=points
+            )
+
+            direction = trajectory.direction  # Returns "Up"
+        """
+        directions = ["Right", "Up", "Left", "Down"]
+        return directions[int((self.bearing / 90) + 0.5) % 4]
+
+    def _raw_direction_index(self) -> int:
+        """
+        Get the cardinal direction index of the trajectory.
+
+        :return int: Cardinal direction index of list ("Right", "Up", "Left", "Down")
+        """
+        return int((self.bearing / 90) + 0.5) % 4
+
+    @computed_field
+    @cached_property
+    def time_interval(self) -> float:
+        """
+        Calculate the time interval of the trajectory.
+
+        :return float: Time interval in seconds
+        """
+        return (self.end - self.start).total_seconds()
+
+    @computed_field
+    @cached_property
+    def geo_location(self) -> GeoLocation:
+        """
+        Get the geographical location of the trajectory as a GeoLocation object.
+
+        :return GeoLocation: GeoLocation object representing the trajectory
+        """
+        geo_location = GeoLocation(type="linestring")
+        for pt in self.points:
+            coordinate = Point()
+            coordinate.point.extend([pt.x, pt.y, pt.z])
+            geo_location.coordinates.append(coordinate)
+        return geo_location
+
+    @computed_field
+    @cached_property
+    def smooth_geo_location(self) -> GeoLocation:
+        """
+        Get the smoothed geographical location of the trajectory as a GeoLocation object.
+
+        :return GeoLocation: GeoLocation object representing the smoothed trajectory
+        """
+        # Create GeoLocation with pre-allocated coordinates list
+        geo_location = GeoLocation(type="linestring")
+        # Use list comprehension to create all points at once
+        geo_location.coordinates = [
+            Point(point=[pt.x, pt.y, pt.z])
+            for pt in self.smooth_trajectory
+        ]
+        return geo_location
 
     def _calculate_distance(self, p1: Coordinate, p2: Coordinate) -> float:
         """
@@ -314,9 +383,124 @@ class Trajectory(TrajectoryBase):
         :param Coordinate p2: The second point
         :return float: Distance between the two points
         """
-        if self.enable_geo:
-            return haversine_distance_coords(p1, p2)
-        return super()._calculate_distance(p1, p2)
+        return euclidean_distance(p1, p2)
+
+    def __str__(self) -> str:
+        """
+        Get a string representation of the trajectory.
+
+        :return str: String representation of the trajectory
+        """
+        return (
+            f"Moving angle {self.bearing}, dir {self.direction:>3}, speed at {self.speed:4.2f}, "
+            f"covered {self.distance:4.2f} in {self.time_interval:4.2f} seconds, id = {self.id}"
+        )
+
+    calibration_type: CalibrationType = CalibrationType.CARTESIAN
+
+
+    @computed_field
+    @cached_property
+    def bearing(self) -> float:
+        """
+        Calculate the bearing of the trajectory from the first to the last point.
+
+        Uses standard Cartesian coordinates (y increasing upward).
+
+        :return float: Bearing in degrees (0-360)
+        """
+        if self.calibration_type == CalibrationType.CARTESIAN:
+            brng = math.atan2(self.last.y - self.head.y, self.last.x - self.head.x) * 180 / math.pi
+            return (brng + 360) % 360
+        if self.calibration_type == CalibrationType.GEO:
+            raise ValueError("Trajectory does not support geographic coordinates")
+        return self._raw_bearing()
+
+    @computed_field
+    @cached_property
+    def speed(self) -> float:
+        """
+        Calculate the average speed of the trajectory.
+
+        Cartesian coordinates are metric, so the base speed is converted to mph. Image coordinates
+        are pixels, so the rate is returned unconverted as pixels per second.
+
+        :return float: Average speed, in mph for cartesian coordinates or pixels/second for image
+        :raises ValueError: If the trajectory is in geographic coordinates
+
+        Example::
+
+            from datetime import datetime, timedelta
+
+            # Create a trajectory that moves 10 meters in 2 seconds
+            points = [
+                Coordinate(x=0, y=0, z=0),
+                Coordinate(x=6, y=8, z=0)  # 10 meters from origin
+            ]
+            start = datetime.now()
+            end = start + timedelta(seconds=2)
+
+            trajectory = Trajectory(
+                id="1",
+                start=start,
+                end=end,
+                points=points
+            )
+
+            # Speed will be converted from m/s to mph
+            avg_speed = trajectory.speed  # Returns approximately 11.18 mph
+        """
+        if self.calibration_type == CalibrationType.CARTESIAN:
+            return self._raw_speed() * MPS_TO_MPH
+        if self.calibration_type == CalibrationType.GEO:
+            raise ValueError("Trajectory does not support geographic coordinates")
+        # Image coordinates are pixels, so the rate is pixels per second and there is nothing
+        # metric to convert -- scaling it by a m/s factor would produce a meaningless number.
+        return self._raw_speed()
+
+    @computed_field
+    @cached_property
+    def speed_over_time(self) -> list[float]:
+        """
+        Calculate the speed of the trajectory over time in segments.
+
+        Uses the same units as :meth:`speed`: mph for cartesian coordinates, pixels per second for
+        image coordinates.
+
+        :return list[float]: Speed for each segment
+        :raises ValueError: If the trajectory is in geographic coordinates
+
+        Example::
+
+            from datetime import datetime, timedelta
+
+            # Create a trajectory with varying speeds
+            points = [
+                Coordinate(x=0, y=0, z=0),
+                Coordinate(x=3, y=4, z=0),  # 5 meters
+                Coordinate(x=6, y=8, z=0)   # 5 more meters
+            ]
+            start = datetime.now()
+            end = start + timedelta(seconds=2)
+
+            trajectory = Trajectory(
+                id="1",
+                start=start,
+                end=end,
+                points=points,
+                speed_segment_size=2
+            )
+
+            # Speeds will be in mph
+            speeds = trajectory.speed_over_time
+            # Returns list of speeds in mph for each segment
+        """
+        if self.calibration_type == CalibrationType.CARTESIAN:
+            return [MPS_TO_MPH * x for x in self._raw_speed_over_time()]
+        if self.calibration_type == CalibrationType.GEO:
+            raise ValueError("Trajectory does not support geographic coordinates")
+        # Pixels per second in image coordinates; see :meth:`speed`.
+        return self._raw_speed_over_time()
 
     def __str__(self) -> str:
         """
@@ -328,3 +512,13 @@ class Trajectory(TrajectoryBase):
             f"Moving angle {self.bearing}, dir {self.direction:>3}, speed at {self.speed:4.2f} mph, "
             f"covered {self.distance:4.2f} meters in {self.time_interval:4.2f} seconds, id = {self.id}"
         )
+
+    @computed_field
+    @cached_property
+    def direction_index(self) -> int:
+        """
+        Cardinal direction index of the trajectory.
+
+        :return int: Index into ("Right", "Up", "Left", "Down")
+        """
+        return self._raw_direction_index()
