+++ services/analytics/behavior-analytics/src/mdx/analytics/core/stream/state/behavior/state_management.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2024-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,254 +14,562 @@
 # limitations under the License.
 
 import logging
-from datetime import datetime
-from typing import Any
-
-import numpy as np
+from dataclasses import dataclass, field
+from datetime import datetime, timedelta, timezone
 
 from mdx.analytics.core.schema.config import AppConfig
-from mdx.analytics.core.schema.models import Behavior, Coordinate, Location, Message, ObjectState
+from mdx.analytics.core.schema.models import Behavior, Message, ObjectState, Coordinate
 from mdx.analytics.core.schema.trajectory.trajectory import Trajectory
-from mdx.analytics.core.transform.calibration.calibration import Calibration
-from mdx.analytics.core.stream.state.behavior.state_management_base import StateMgmtBase
-from mdx.analytics.core.utils.crs import CoordinateReferenceSystem
-from mdx.analytics.core.utils.distance_util import haversine_distance
-from mdx.analytics.core.utils.schema_util import model_to_embeddings
+from mdx.analytics.core.stream.state.behavior import point_sampling
+from mdx.analytics.core.stream.state.behavior.behavior_holdback import BehaviorHoldback
+from mdx.analytics.core.transform.calibration.calibration_base import CalibrationBase
+from mdx.analytics.core.utils.crp import CRP
+from mdx.analytics.core.utils.schema_util import get_sensor_id_from_behavior_id, model_to_embeddings
 
 logger = logging.getLogger(__name__)
 
 
-class StateMgmt(StateMgmtBase):
+
+@dataclass
+class BehaviorBatch:
     """
-    Module to manage behavior state in euclidean coordinate systems.
+    Everything one batch of messages produced.
 
-    This class extends StateMgmtBase to provide specific implementations for:
-    - Map matching and road network integration
-    - Edge-based speed calculations
-    - Trajectory smoothing and subsampling
-    - Behavior state updates with map context
+    A plain dataclass rather than a pydantic model: this never crosses a wire, and it is built once
+    per batch on the hot path. The lists hold live references, so enriching an entry of
+    ``active_behaviors`` in place -- adding anomaly info, compacting -- is visible to whatever
+    is written. Replacing the objects instead of mutating them would break that.
 
-    :ivar CoordinateReferenceSystem | None crs: Reference system used for map matching, required by
-        ``mapMatchingClasses`` object types. Fixed for the lifetime of the app, so it is a constructor
-        dependency rather than a per-call argument.
-    :ivar int mapmatching_success_cnt: Counter for successful map matching operations.
-    :ivar int mapmatching_total_cnt: Total counter for map matching operations.
+    :ivar list[Behavior] active_behaviors: One per track that this batch updated. Feed these to
+        detection and enrich them in place.
+    :ivar list[Behavior] trip_behaviors: One per track, covering the previous tail plus this batch's
+        points. Used for tripwire and ROI events; always empty in geographic coordinates, which does
+        not track trips.
+    :ivar list[Behavior] behaviors_to_write: What to write to the behavior stream. The emission
+        policy is already resolved here, so no caller needs to know whether ``behaviorEmitOnce`` is
+        on: it is either the active behaviors or the tracks that just ended.
 
     Examples::
-        >>> state_manager = StateMgmt(config, calibration, crs)
         >>> batch = state_manager.process_batch(messages_map)
-        >>> print(f"Writing {len(batch.behaviors_to_write)} behavior(s)")
+        >>> app.write_behaviors(batch.behaviors_to_write)
     """
 
-    def __init__(
-        self, config: AppConfig, calibration: Calibration, crs: CoordinateReferenceSystem | None
-    ) -> None:
-        super().__init__(config=config, calibration=calibration)
-        self.crs = crs
-        # for debug
-        self.mapmatching_success_cnt = 0
-        self.mapmatching_total_cnt = 0
+    active_behaviors: list[Behavior] = field(default_factory=list)
+    trip_behaviors: list[Behavior] = field(default_factory=list)
+    behaviors_to_write: list[Behavior] = field(default_factory=list)
 
-    def _update_object_state_model(self, state: ObjectState, embeddings: list[list[float]]) -> None:
-        """No-op for geographic coordinates - model with embeddings updates not needed."""
-        return
 
-    def _create_trajectory(self, id: str, start: datetime, end: datetime, 
+class StateMgmt:
+    """
+    Base class for object tracking over a period of time, used for metadata in protobuf.
+
+    This class provides core functionality for:
+    - Object state management and tracking
+    - Behavior state updates and maintenance
+    - Sensor timestamp management
+    - Object type scoring and clustering
+    - Trajectory and behavior generation
+
+    :ivar AppConfig config: Configuration object for the application.
+    :ivar CalibrationBase calibration: Calibration object for the application.
+    :ivar dict[str, ObjectState] state: Dictionary to store object states.
+    :ivar dict[str, datetime] sensor_latest_timestamp: Dictionary mapping sensor IDs to their latest timestamps.
+    :ivar BehaviorHoldback behavior_holdback: Holds behaviors back when ``behaviorEmitOnce`` is enabled.
+
+    Examples::
+        >>> state_manager = StateMgmt(config, calibration)
+        >>> batch = state_manager.process_batch(messages_map)
+        >>> print(f"Written {len(batch.behaviors_to_write)} behavior(s)")
+    """
+
+    def __init__(self, config: AppConfig, calibration: CalibrationBase) -> None:
+        self.config: AppConfig = config
+        self.calibration: CalibrationBase = calibration
+        self.state: dict[str, ObjectState] = dict()
+        self.sensor_latest_timestamp: dict[str, datetime] = dict()
+        self.behavior_holdback: BehaviorHoldback = BehaviorHoldback()
+
+    def process_batch(self, messages_map: dict[str, list[Message]]) -> BehaviorBatch:
+        """
+        Process one whole batch of messages, grouped by track key.
+
+        Taking the entire batch rather than a key at a time is deliberate. Deciding which tracks have
+        fallen silent is only sound once every message in the batch has reached its state and every
+        sensor clock has advanced; judging it key by key would end tracks whose own messages are
+        still sitting later in the same batch. Owning the loop here makes that impossible to get
+        wrong from the outside -- one call is one batch.
+
+        :param dict[str, list[Message]] messages_map: Messages grouped by key (sensor ID + object ID),
+            as produced by ``messages_to_map``.
+        :return BehaviorBatch: Behaviors updated by this batch, their trip states, and what to write.
+
+        Examples::
+            >>> batch = state_manager.process_batch(messages_to_map(messages))
+            >>> app.write_behaviors(batch.behaviors_to_write)
+        """
+        batch = BehaviorBatch()
+
+        for message_key, messages in messages_map.items():
+            behavior, trip_behavior = self._process_key(message_key, messages)
+            if behavior:
+                batch.active_behaviors.append(behavior)
+            if trip_behavior:
+                batch.trip_behaviors.append(trip_behavior)
+
+        # Retain before ending, so a track that both produced a behavior and fell silent in the same
+        # batch still reaches the stream: the sweep below releases whatever was just retained.
+        if self.config.behavior_emit_once:
+            self.behavior_holdback.retain(batch.active_behaviors)
+
+        # Runs in both modes -- ending a track is what reclaims its state, so per-batch mode depends
+        # on it too, even though it holds nothing back.
+        self._end_inactive_behaviors()
+
+        if self.config.behavior_emit_once:
+            batch.behaviors_to_write = self.behavior_holdback.take_ended()
+        else:
+            batch.behaviors_to_write = (
+                self._carry_over_held_behaviors(batch.active_behaviors) + batch.active_behaviors
+            )
+
+        return batch
+
+    def _carry_over_held_behaviors(self, active_behaviors: list[Behavior]) -> list[Behavior]:
+        """
+        Hand over anything still held back, once, after ``behaviorEmitOnce`` is switched off.
+
+        ``behaviorEmitOnce`` is runtime-updatable, so tracks can be left held when it flips. They
+        cannot simply be dropped -- a track that fell silent around the flip would never be written
+        at all -- but neither can they trickle out as they end: per-batch output resumes immediately
+        and writes progressively more complete behaviors for the same IDs, so a snapshot released
+        later would arrive with an older ``end`` than one already sent and regress it downstream.
+
+        So the holdback is emptied in one go, minus any track already producing fresh output this
+        batch. After that it stays empty and this is a no-op.
+
+        :param list[Behavior] active_behaviors: Behaviors this batch produced, whose tracks already
+            have fresher output and so need nothing carried over.
+        :return list[Behavior]: Held behaviors worth writing, ahead of this batch's own output.
+        """
+        if not self.behavior_holdback.pending and not self.behavior_holdback.ended:
+            return []
+
+        fresh_ids = {behavior.id for behavior in active_behaviors}
+        carried = [behavior for behavior in self.behavior_holdback.flush() if behavior.id not in fresh_ids]
+        logger.info(f"behaviorEmitOnce switched off; handing over {len(carried)} held behavior(s)")
+
+        return carried
+
+    def flush_behaviors(self) -> list[Behavior]:
+        """
+        Return every behavior still held back, ended or not, and drop it.
+
+        Intended for shutdown, where tracks that were still live would otherwise be lost. Returns an
+        empty list unless emit-once is enabled, since nothing is ever held back in per-batch mode.
+
+        :return list[Behavior]: Behaviors to write before the app stops.
+
+        Examples::
+            >>> app.write_behaviors(state_manager.flush_behaviors())
+        """
+        return self.behavior_holdback.flush()
+
+    def live_object_ids(self) -> list[str]:
+        """
+        Keys of the tracks currently held in state, for detectors that need to know what is live.
+
+        :return list[str]: Track keys (sensor ID + object ID).
+
+        Examples::
+            >>> detector.update_live_object(state_manager.live_object_ids())
+        """
+        return list(self.state.keys())
+
+    def _get_current_timestamp(self, sensorId: str) -> datetime | None:
+        """
+        Get the current timestamp for a sensor.
+
+        In simulation mode, returns the latest timestamp for the sensor.
+        Otherwise, returns the current UTC time.
+
+        :param str sensorId: The sensor ID to get timestamp for.
+        :return datetime | None: Current timestamp for the sensor, or None if not found in simulation mode.
+        :raises ValueError: If in simulation mode and no timestamp exists for sensor.
+
+        Examples::
+            >>> state_manager = StateMgmt(config)
+            >>> timestamp = state_manager._get_current_timestamp("sensor1")
+            >>> print(f"Current timestamp: {timestamp}")
+        """
+        if not self.config.in_simulation_mode:
+            return datetime.now(timezone.utc)
+        return self.sensor_latest_timestamp.get(sensorId)
+
+    def _update_sensor_latest_timestamp(self, messages: list[Message]) -> None:
+        """
+        Update the latest timestamp for each sensor based on incoming messages.
+
+        :param list[Message] messages: List of messages containing sensor timestamps.
+        :return: None
+
+        Examples::
+            >>> state_manager = StateMgmt(config)
+            >>> messages = [Message(sensor=Sensor(id="sensor1"), timestamp=datetime.now())]
+            >>> state_manager._update_sensor_latest_timestamp(messages)
+        """
+        for msg in messages:
+            if (msg.sensor.id not in self.sensor_latest_timestamp) or (
+                msg.timestamp > self.sensor_latest_timestamp[msg.sensor.id]
+            ):
+                logger.info(f"Updating sensor latest timestamp: {msg.sensor.id} to {msg.timestamp}")
+                self.sensor_latest_timestamp[msg.sensor.id] = msg.timestamp
+
+    def _end_inactive_behaviors(self) -> None:
+        """
+        End tracks that have been quiet for at least ``behaviorStateValidInterval`` seconds.
+
+        That gap is the same threshold :meth:`_is_valid_state` uses to reject a continuation, so once
+        it passes the track can no longer be extended -- any later observation starts a new one, which
+        makes it the point at which the track is provably over. Ending is therefore also when the
+        object state is reclaimed: nothing further can be added to it, and a returning object ID
+        simply starts a fresh track.
+
+        Inactivity is measured per sensor in *event* time (``sensor_latest_timestamp``), not against
+        the wall clock, for two reasons. A sensor's clock must only advance on its own data, so busy
+        sensors cannot age out the tracks of a sensor whose batch is still in flight. And because the
+        threshold is short, wall-clock comparison would misread ingestion lag or backpressure as
+        silence and cut live tracks short. This mirrors :meth:`_is_valid_state`, which also compares
+        event timestamps.
+
+        A sensor that stops streaming freezes its own clock, so its tracks are never ended and their
+        state is retained. That residue is bounded -- a silent sensor contributes no new object IDs --
+        and it is released when the app stops.
+
+        Only correct at end of batch, once every key has been through :meth:`_process_key` and all
+        sensor clocks are current -- see :meth:`process_batch`, its only caller.
+
+        :return: None
+        """
+        for behavior_id in list(self.state.keys()):
+            sensor_timestamp = self.sensor_latest_timestamp.get(get_sensor_id_from_behavior_id(behavior_id))
+            if not sensor_timestamp:
+                continue
+            if (sensor_timestamp - self.state[behavior_id].end).total_seconds() >= self.config.behavior_state_valid_interval:
+                logger.info(f"Track ended, releasing state: {behavior_id}")
+                del self.state[behavior_id]
+                self.behavior_holdback.end_track(behavior_id, reason="track inactive")
+
+    def _is_valid_state(self, old_state: ObjectState, new_state: ObjectState, interval: int) -> bool:
+        """
+        Check if the state transition is valid based on the time interval.
+
+        :param ObjectState old_state: Old state stored in memory.
+        :param ObjectState new_state: New generated object state.
+        :param int interval: Maximum allowed interval in seconds.
+        :return bool: True if the state transition is valid, False otherwise.
+
+        Examples::
+            >>> state_manager = StateMgmt(config)
+            >>> old_state = ObjectState(end=datetime.now())
+            >>> new_state = ObjectState(start=datetime.now() + timedelta(seconds=2))
+            >>> is_valid = state_manager._is_valid_state(old_state, new_state)
+            >>> print(f"State transition valid: {is_valid}")
+        """
+        valid = (new_state.start - old_state.end).total_seconds() < interval and new_state.start >= old_state.end
+        if not valid:
+            logger.info(
+                f"invalid old state, id: {old_state.id}, old state end: {old_state.end}, new state start: {new_state.start}"
+            )
+        return valid
+
+    def _create_trajectory(self, id: str, start: datetime, end: datetime,
                           points: list[Coordinate]) -> Trajectory:
-        """Returns IncrementalTrajectory with geographic/direction configuration."""
+        """
+        Build the trajectory for a track, in cartesian or image coordinates.
+
+        ``Trajectory`` gates bearing and speed on the calibration type, which is read here rather
+        than captured, so a calibration switch is picked up on the next batch. Geographic
+        coordinates need map matching and a different distance metric -- see :class:`StateMgmtG`.
+
+        :param str id: Track key the trajectory belongs to.
+        :param datetime start: Start of the trajectory.
+        :param datetime end: End of the trajectory.
+        :param list[Coordinate] points: Accumulated points.
+        :return Trajectory: Trajectory for this coordinate system.
+        """
         return Trajectory(
             id=id,
             start=start,
             end=end,
             points=points,
-            enable_geo=self.config.traj_geo_coord_enable,
             smooth_min_points=self.config.traj_smooth_min_points,
             smooth_window_size=self.config.traj_smooth_window_size,
             distance_stride=self.config.traj_distance_stride,
             speed_segment_size=self.config.traj_speed_segment_size,
-            direction_mode=self.config.traj_direction_mode,
-            direction_based_cluster_mode=self.config.traj_direction_cluster_mode,
+            calibration_type=self.calibration.calibration_type,
         )
 
-    def _process_key(self, message_key: str, messages: list[Message]) -> tuple[Behavior | None, None]:
+    def _update_object_state_model(self, state: ObjectState, embeddings: list[list[float]]) -> None:
+        """Update the clustering model of an object state. Override in child classes if needed."""
+        if state.model:
+            clustering_model = CRP().update_model(state.model, embeddings, self.config.cluster_threshold)
+        else:
+            clustering_model = CRP().cluster(embeddings, self.config.cluster_threshold)
+        state.model = clustering_model
+    
+    def _get_object_state_and_message(
+        self, message_key: str, messages: list[Message]
+    ) -> tuple[ObjectState | None, Message | None]:
         """
-        Advance one track, in geographic coordinates.
+        Get new state of an object and last message.
 
-        Differs from the base in two ways: the behavior is enriched with map-matched road edges and
-        speed over them, and no trip state is tracked, so the trip behavior is always ``None``.
+        This method processes messages to create or update an object state, including:
+        - Filtering messages by time threshold
+        - Extracting coordinates and embeddings
+        - Computing object type scores
+        - Creating or updating clustering model
+        - Managing state transitions
 
         :param str message_key: Key for the message (sensor ID + object ID).
-        :param list[Message] messages: Messages for this key in the current batch.
-        :return tuple[Behavior | None, None]: The track's behavior, or ``(None, None)`` when the
-            messages yielded no valid state.
+        :param list[Message] messages: List of messages to process.
+        :return tuple[ObjectState | None, Message | None]: Tuple containing the object state and last message, or (None, None) if invalid.
+
+        Examples::
+            >>> state_manager = StateMgmt(config)
+            >>> messages = [Message(sensor=Sensor(id="sensor1"), timestamp=datetime.now())]
+            >>> state, msg = state_manager._get_object_state_and_message("sensor1_obj1", messages)
+            >>> print(f"Created state with {len(state.points)} points")
         """
-        self._update_sensor_latest_timestamp(messages)
-        state, last_message = self._get_object_state_and_message(message_key, messages)
+        state, _, last_message = self._get_object_trip_state_and_message(message_key, messages)
+        return state, last_message
 
-        # When getting no new behavior
-        if not state or not last_message:
-            return None, None
+    def _get_object_trip_state_and_message(
+        self, message_key: str, messages: list[Message]
+    ) -> tuple[ObjectState | None, ObjectState | None, Message | None]:
+        """
+        Get new state of an object, trip state and last message.
 
-        behavior_traj = self._create_trajectory(state.id, state.start, state.end, state.points)
-        return self._get_behavior(state, behavior_traj, last_message), None
+        This method processes messages to create or update object and trip states, including:
+        - Filtering messages by time threshold
+        - Computing trip states with minimum points
+        - Managing state transitions and sampling
+        - Updating clustering models
+
+        :param str message_key: Key for the message (sensor ID + object ID).
+        :param list[Message] messages: List of messages to process.
+        :return tuple[ObjectState | None, ObjectState | None, Message | None]: Tuple containing the object state, trip state and last message, or (None, None, None) if invalid.
+
+        Examples::
+            >>> state_manager = StateMgmt(config)
+            >>> messages = [Message(sensor=Sensor(id="sensor1"), timestamp=datetime.now())]
+            >>> state, trip_state, msg = state_manager._get_object_trip_state_and_message("sensor1_obj1", messages)
+            >>> print(f"Created states with {len(state.points)} points")
+        """
+        # Skip invalid or dummy messages
+        if not message_key or not messages or message_key.endswith("dummy"):
+            return None, None, None
+
+        # Configure trip tracking parameters
+        sensor_id = messages[0].sensor.id
+        tripwire_min_points = self.config.sensor_tripwire_min_points(sensor_id)
+        min_trip_length = tripwire_min_points * 2
+        min_trip_length_minus_one = min_trip_length - 1
+
+        # Filter messages in stages
+        sorted_messages = sorted(list(messages), key=lambda x: x.timestamp)
+        time_threshold = sorted_messages[-1].timestamp - timedelta(seconds=self.config.behavior_water_mark)
+        state = self.state.get(message_key)
+
+        # 1) Drop messages outside time window or before global behavior threshold
+        filtered_by_time = [
+            msg
+            for msg in sorted_messages
+            if msg.timestamp >= time_threshold and msg.timestamp > self.config.behavior_time_threshold
+        ]
+        dropped_by_time = len(sorted_messages) - len(filtered_by_time)
+        if dropped_by_time:
+            logger.warning(
+                f"{dropped_by_time} message(s) filtered out (older than {self.config.behavior_water_mark}s window or "
+                f"before behavior_time_threshold {self.config.behavior_time_threshold}) for {message_key}"
+            )
+
+        # 2) Widened cutoff + split into in-order / in-tolerance in a single pass.
+        in_order_msgs: list[Message] = []
+        in_tolerance_msgs: list[Message] = []
+        if state is not None:
+            cutoff = state.end - timedelta(seconds=self.config.behavior_state_end_tolerance_sec)
+            dropped = 0
+            for msg in filtered_by_time:
+                if msg.timestamp > state.end:
+                    in_order_msgs.append(msg)
+                elif msg.timestamp > cutoff:
+                    in_tolerance_msgs.append(msg)
+                else:
+                    dropped += 1
+            if dropped:
+                logger.warning(
+                    f"{dropped} message(s) filtered out (older than {cutoff}, "
+                    f"tolerance {self.config.behavior_state_end_tolerance_sec}s) for {message_key}"
+                )
+        else:
+            in_order_msgs = filtered_by_time
+
+        if not in_order_msgs:
+            if in_tolerance_msgs:
+                logger.debug(
+                    f"Tolerance-only batch for {message_key}: "
+                    f"{len(in_tolerance_msgs)} late message(s) dropped"
+                )
+            return None, None, None
+
+        coordinates = [msg.object.coordinate for msg in in_order_msgs]
+        # Per-frame bboxes kept aligned 1:1 with ``coordinates`` (and thus ``points``).
+        bboxes = [msg.object.bbox for msg in in_order_msgs]
+        embeddings = [
+            msg.object.embedding.vector
+            for msg in (in_order_msgs + in_tolerance_msgs)
+            if msg.object.confidence >= self.config.object_confidence_threshold
+            and msg.object.embedding and msg.object.embedding.vector
+        ]
+        last_x_points = coordinates[-min_trip_length_minus_one:]
+        last_x_bboxes = bboxes[-min_trip_length_minus_one:]
+
+        new_state = ObjectState(
+            id=message_key,
+            start=in_order_msgs[0].timestamp,
+            end=in_order_msgs[-1].timestamp,
+            points=coordinates,
+            bboxes=bboxes,
+            lastXpoints=last_x_points,
+            lastXbboxes=last_x_bboxes,
+            tail_ts=[m.timestamp for m in in_order_msgs[-point_sampling.TAIL_CAP:]],
+        )
+
+        # If no old state or invalid transition, use new state for both
+        if not state or not self._is_valid_state(state, new_state, self.config.behavior_state_valid_interval):
+            if state:
+                # The gap proves the previous track ended, so its retained behavior is final; ending it
+                # here -- before this batch retains the replacement under the same key -- keeps at most
+                # one pending behavior per key.
+                self.behavior_holdback.end_track(message_key, reason="track replaced after discontinuity")
+            self.state[message_key] = new_state
+            self._update_object_state_model(new_state, embeddings)
+            logger.info(f"Created new Object State: {message_key}\n"
+                        f"  Start: {new_state.start}\n"
+                        f"  End: {new_state.end}\n"
+                        f"  Points: {len(new_state.points)}\n"
+                        f"  TimeInterval: {new_state.end - new_state.start}\n"
+                        f"  Length: {len(new_state.points)}")
+            return new_state, new_state, in_order_msgs[-1]
+
+        # === Update existing state ===
+        
+        # Prepare trip data (combination of old and new)
+        trip_points = state.lastXpoints + new_state.points
+        trip_bboxes = state.lastXbboxes + new_state.bboxes
+        if new_state.time_interval != 0 and len(new_state.points) > 1:
+            interval = new_state.time_interval / (len(new_state.points) - 1)
+        else:
+            interval = (new_state.start - state.end).total_seconds()
+        trip_start = new_state.start - timedelta(seconds=len(state.lastXpoints) * interval)
+
+        point_sampling.append_sampled(state, in_order_msgs)
+        point_sampling.insert_tolerance_messages(state, in_tolerance_msgs, message_key)
+        point_sampling.halve_if_needed(state, self.config.behavior_max_points)
+
+        state.end = new_state.end
+        state.lastXpoints = trip_points[-min_trip_length_minus_one:]
+        state.lastXbboxes = trip_bboxes[-min_trip_length_minus_one:]
+        self._update_object_state_model(state, embeddings)
+
+        # Create trip state
+        trip_state = ObjectState(
+            id=message_key,
+            start=trip_start,
+            end=new_state.end,
+            points=trip_points,
+            bboxes=trip_bboxes,
+        )
+
+        self.state[message_key] = state
+        logger.info(f"Updated Object State: {message_key}\n"
+                    f"  Start: {state.start}\n"
+                    f"  End: {state.end}\n"
+                    f"  Points: {len(state.points)}\n"
+                    f"  TimeInterval: {state.end - state.start}\n"
+                    f"  Length: {len(state.points)}")
+        return state, trip_state, in_order_msgs[-1]
 
     def _get_behavior(self, state: ObjectState, tr: Trajectory, message: Message) -> Behavior:
         """
         Get behavior from object state, trajectory and message.
 
-        This method:
-        1. Creates info dictionary with cluster information
-        2. Performs map matching if object type is configured for it
-        3. Calculates speed over edges for matched trajectories
-        4. Constructs and returns behavior object with all attributes
-
-        :param ObjectState state: Updated object state containing points and metadata.
-        :param Trajectory tr: Processed trajectory with smoothing and subsampling.
-        :param Message message: Last message containing sensor and object information.
-        :return Behavior: Behavior object containing all processed information.
+        :param ObjectState state: Updated object state.
+        :param Trajectory tr: Trajectory of the behavior.
+        :param Message message: Last message containing object information.
+        :return Behavior: Updated behavior object.
 
         Examples::
-            >>> state = ObjectState(id="obj1", points=[...])
-            >>> tr = Trajectory(id="obj1", points=[...])
+            >>> state_manager = StateMgmt(config)
+            >>> state = ObjectState(points=[...])
+            >>> trajectory = Trajectory(direction_index=1)
             >>> message = Message(sensor=Sensor(id="sensor1"))
-            >>> behavior = state_manager._get_behavior(state, tr, message)
-            >>> print(f"Created behavior with {len(behavior.locations)} locations")
+            >>> behavior = state_manager._get_behavior(state, trajectory, message)
+            >>> print(f"Created behavior with direction {behavior.direction}")
         """
-        # get edges and speed over edge
-        edges = []
-        obj_type = message.object.type
-        if obj_type in self.config.map_matching_classes:
-            trajectory = self._subsample_for_mapmatching(tr.smooth_trajectory, self.config.map_matching_max_points)
-            trajectory = [Location(lat=coord.y, lon=coord.x) for coord in trajectory]
-            _, _, matched_lattice = self.crs.road_network.map_matching_for_edge(trajectory)
-            edges, _ = self._speed_over_edge(tr, matched_lattice)
-            # speed_over_time = speed_over_edge  # choose between speed_over_edge or speed_over_time since for now there is no extra field for speed over edge
-            # snapped trajectory (not used in behavior output atm)
 
         return Behavior(
-            id=tr.id,
-            timestamp=tr.start,
-            end=tr.end,
+            id=state.id,
+            timestamp=state.start,
+            end=state.end,
+            timeInterval=state.time_interval,
+            embeddings=model_to_embeddings(state.model),
             locations=tr.geo_location,
             locationsBboxes=state.bboxes,
             smoothLocations=tr.smooth_geo_location,
             distance=tr.distance,
             speed=tr.speed,
             speedOverTime=tr.speed_over_time,
-            timeInterval=tr.time_interval,
             bearing=tr.bearing,
             direction=tr.direction,
-            length=len(state.points),
+            length=len(tr.points),
             place=message.place,
             sensor=message.sensor,
             object=message.object,
             event=message.event,
             videoPath=message.videoPath,
-            embeddings=model_to_embeddings(state.model),
-            edges=edges,
-            info=(
-                {}
-                if self.config.inference.enable
-                else {
-                    "cluster.modelVersion": "directionBasedModel",
-                    "cluster.index": str(tr.direction_based_cluster_index),
-                }
-            ),
+            info={
+                "cluster.modelVersion": "directionBasedModel",
+                "cluster.index": str(tr.direction_index)
+            }
         )
 
-    def _subsample_for_mapmatching(self, points: list[Any], target_point: int) -> list[Any]:
+    def _process_key(self, message_key: str, messages: list[Message]) -> tuple[Behavior | None, Behavior | None]:
         """
-        Subsample points for map matching to reduce computational complexity.
+        Advance one track with its share of the current batch.
 
-        This method:
-        1. Returns original points if length is less than target or target is invalid
-        2. Otherwise, evenly samples points to match target count
+        Updates the sensor clock, folds the messages into the track's object and trip state, and
+        builds a behavior for each. Called only by :meth:`process_batch`, which owns the batch-wide
+        steps -- expiry and the emission policy -- that are only sound once every key is done.
 
-        :param list[Any] points: List of points to subsample.
-        :param int target_point: Target number of points after subsampling.
-        :return list[Any]: List of subsampled points.
+        Subclasses override this when a coordinate system produces something different; geographic
+        coordinates, for instance, track no trips and return ``None`` for the trip behavior.
 
-        Examples::
-            >>> points = [Point(x=1, y=1), Point(x=2, y=2), Point(x=3, y=3)]
-            >>> subsampled = state_manager._subsample_for_mapmatching(points, 2)
-            >>> print(f"Subsampled to {len(subsampled)} points")
+        :param str message_key: Key for the message (sensor ID + object ID).
+        :param list[Message] messages: Messages for this key in the current batch.
+        :return tuple[Behavior | None, Behavior | None]: The track's behavior and trip behavior, or
+            ``(None, None)`` when the messages yielded no valid state.
         """
-        points_subsample = []
-        if (len(points) <= target_point) or (target_point <= 0):
-            points_subsample = points
-        else:
-            points_subsample = [
-                points[int(i)] for i in np.linspace(0, len(points) - 1, num=target_point, endpoint=True)
-            ]
+        self._update_sensor_latest_timestamp(messages)
+        state, trip_state, last_message = self._get_object_trip_state_and_message(message_key, messages)
+        if not state or not trip_state or not last_message:
+            return None, None
 
-        return points_subsample
+        # Build trajectories
+        behavior_traj = self._create_trajectory(state.id, state.start, state.end, state.points)
+        trip_traj = self._create_trajectory(trip_state.id, trip_state.start, trip_state.end, trip_state.points)
 
-    def _speed_over_edge(self, tr: Trajectory, matched_lattice) -> tuple[list[str], list[float]]:
-        """
-        Calculate speed over matched road network edges.
+        behaviorMessage = self._get_behavior(state, behavior_traj, last_message)
+        tripMessage = self._get_behavior(trip_state, trip_traj, last_message)
 
-        This method:
-        1. Maps trajectory points to road network edges
-        2. Calculates speed for each edge segment
-        3. Converts speeds from km/h to mph
-        4. Returns edge IDs and corresponding speeds
-
-        :param Trajectory tr: Processed trajectory with smoothing.
-        :param matched_lattice: Matched road network lattice from map matching.
-        :return tuple[list[str], list[float]]: Tuple containing list of edge IDs and corresponding speeds.
-
-        Examples::
-            >>> tr = Trajectory(id="obj1", points=[...])
-            >>> edges, speeds = state_manager._speed_over_edge(tr, matched_lattice)
-            >>> print(f"Calculated speeds for {len(edges)} edges")
-        """
-        kmph_to_mph = 0.621371
-        map_edge_to_traj = {}
-
-        for idx, m in enumerate(matched_lattice):
-            edge_id = m.edge_m.label
-            # edge_o_p1_loc = Location(lat=m.edge_o.p1[0],lon=m.edge_o.p1[1])
-            # traj_idx = tr.smooth_geo_trajectory.index(edge_o_p1_loc)
-            edge_o_p1_coord = Coordinate(x=m.edge_o.p1[1], y=m.edge_o.p1[0])
-            traj_idx = tr.smooth_trajectory.index(edge_o_p1_coord)
-            if not m.edge_o.is_point():
-                # edge_o_p2_loc = Location(lat=m.edge_o.p2[0],lon=m.edge_o.p2[1])
-                # traj_idx = tr.smooth_geo_trajectory.index(edge_o_p2_loc)
-                edge_o_p2_coord = Coordinate(x=m.edge_o.p2[1], y=m.edge_o.p2[0])
-                traj_idx = tr.smooth_trajectory.index(edge_o_p2_coord)
-            if edge_id in map_edge_to_traj:
-                map_edge_to_traj[edge_id].append(traj_idx)
-            else:
-                map_edge_to_traj[edge_id] = [traj_idx]
-
-        edges = []
-        speed_over_edge = []
-        for edge_id, traj_point_idx in map_edge_to_traj.items():
-            edges.append(edge_id)
-            # If the trajectory has only one point, skip the speed calculation for this edge
-            if len(tr.smooth_trajectory) > 1:
-                from_idx = max(0, traj_point_idx[0] - 1)
-                to_idx = traj_point_idx[-1]
-                if from_idx == to_idx:
-                    to_idx += 1
-                slice = tr.smooth_trajectory[from_idx : to_idx + 1]
-                interval = (tr.end - tr.start).total_seconds() * (len(slice) - 1) / (len(tr.smooth_trajectory) - 1)
-
-                # If interval is 0, skip the speed calculation for this edge
-                if interval == 0:
-                    continue
-
-                speed = (
-                    kmph_to_mph
-                    * sum(
-                        haversine_distance(
-                            Location(lat=slice[i].y, lon=slice[i].x), Location(lat=slice[i + 1].y, lon=slice[i + 1].x)
-                        )
-                        for i in range(len(slice) - 1)
-                    )
-                    * 3.6
-                    / interval
-                )
-                speed_over_edge.append(round(speed, 2))
-
-        return edges, speed_over_edge
+        return behaviorMessage, tripMessage
